# backend/docs/openapi.yaml
openapi: 3.0.3
info:
  title: GreenHerb API
  description: |
    API REST para a gestão de planos de cultivo da estufa GreenHerb.
    Projeto académico — Sprint 2.
  version: 0.1.0
  contact:
    name: Equipa GreenHerb
servers:
  - url: http://localhost:5000
    description: Servidor de desenvolvimento local
...
paths:
...
  /api/lotes:
    get:
      tags: [Lotes]
      summary: Listar lotes
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Lista de lotes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/LoteCultivo'
        '401': { $ref: '#/components/responses/Unauthorized' }
    post:
      tags: [Lotes]
      summary: Criar lote
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoteCultivo'
      responses:
        '201':
          description: Lote criado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoteCultivo'
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /api/lotes/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: string }
    get:
      tags: [Lotes]
      summary: Detalhe do lote
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Lote
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoteCultivo'
        '404': { $ref: '#/components/responses/NotFound' }
    put:
      tags: [Lotes]
      summary: Editar lote (criador ou Administrador)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoteCultivo'
      responses:
        '200':
          description: Lote atualizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoteCultivo'
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
    delete:
      tags: [Lotes]
      summary: Eliminar lote (Administrador)
      security:
        - bearerAuth: []
      responses:
        '204': { description: Eliminado }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /api/lotes/{id}/perda:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: string }
    post:
      tags: [Lotes]
      summary: Registar perda no lote
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [quantidade, motivo]
              properties:
                quantidade: { type: integer, minimum: 1 }
                motivo: { type: string }
      responses:
        '200':
          description: Lote atualizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoteCultivo'
        '400': { $ref: '#/components/responses/BadRequest' }
        '404': { $ref: '#/components/responses/NotFound' }
  /api/lotes/{id}/estado:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: string }
    put:
      tags: [Lotes]
      summary: Mudar estado do lote
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [estado]
              properties:
                estado:
                  type: string
                  enum: [ativo, concluído, comprometido]
      responses:
        '200':
          description: Lote atualizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoteCultivo'
...
  /api/medicoes:
    get:
      tags: [Medições]
      summary: Listar medições
      security:
        - bearerAuth: []
      parameters:
        - name: loteId
          in: query
          schema: { type: string }
      responses:
        '200':
          description: Lista de medições
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/MedicaoAmbiental'
    post:
      tags: [Medições]
      summary: Registar medição (gera alertas automáticos)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MedicaoAmbiental'
      responses:
        '201':
          description: Medição registada com lista de alertas gerados
          content:
            application/json:
              schema:
                type: object
                properties:
                  medicao:
                    $ref: '#/components/schemas/MedicaoAmbiental'
                  alertas:
                    type: array
                    items:
                      $ref: '#/components/schemas/Alerta'
  /api/medicoes/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: string }
    get:
      tags: [Medições]
      summary: Detalhe da medição
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Medição
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MedicaoAmbiental'
    delete:
      tags: [Medições]
      summary: Eliminar medição (Administrador)
      security:
        - bearerAuth: []
      responses:
        '204': { description: Eliminada }
...
  /api/planos:
    get:
      tags: [Planos]
      summary: Listar planos
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Lista de planos
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PlanoCultivo'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
    post:
      tags: [Planos]
      summary: Criar plano
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PlanoCultivoInput'
      responses:
        '201':
          description: Plano criado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlanoCultivo'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/planos/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          example: 65a1b2c3d4e5f60718293a4b
    get:
      tags: [Planos]
      summary: Detalhe de um plano
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Plano encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlanoCultivo'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
    put:
      tags: [Planos]
      summary: Editar plano (criador ou Administrador)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PlanoCultivoInput'
      responses:
        '200':
          description: Plano atualizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlanoCultivo'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
    delete:
      tags: [Planos]
      summary: Eliminar plano (apenas Administrador)
      security:
        - bearerAuth: []
      responses:
        '204':
          description: Plano eliminado
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    BadRequest:
      description: Pedido inválido
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Autenticação requerida ou token inválido
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Forbidden:
      description: Sem permissões para a operação
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Recurso não encontrado
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    InternalError:
      description: Erro interno do servidor
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
    Error:
      type: object
      properties:
        erro:
          type: string
          example: Mensagem descritiva do erro.
    Utilizador:
      type: object
      properties:
        _id:
          type: string
          example: 65a1b2c3d4e5f60718293a4b
        nome:
          type: string
          example: Ana Administradora
        email:
          type: string
          format: email
          example: admin@greenherb.local
        perfil:
          type: string
          enum: [Técnico, Responsável, Administrador]
          example: Administrador
        ativo:
          type: boolean
          example: true
        dataCriacao:
          type: string
          format: date-time
...
    DadosRegular:
      type: object
      required:
        - tempMin
        - tempMax
        - humidadeMin
        - humidadeMax
        - luminosidadeMin
        - luminosidadeMax
        - duracaoCiclo
        - frequenciaRega
        - volumeRega
        - frequenciaFertilizacao
      properties:
        tempMin: { type: number }
        tempMax: { type: number }
        humidadeMin: { type: number, minimum: 0, maximum: 100 }
        humidadeMax: { type: number, minimum: 0, maximum: 100 }
        luminosidadeMin: { type: number, minimum: 0 }
        luminosidadeMax: { type: number, minimum: 0 }
        duracaoCiclo: { type: integer, minimum: 1 }
        frequenciaRega: { type: integer, minimum: 1 }
        volumeRega: { type: number, minimum: 0 }
        frequenciaFertilizacao: { type: integer, minimum: 1 }
    DadosEmergencia:
      type: object
      required: [tipoIntervencao, intervaloMinimo, dosagem]
      properties:
        tipoIntervencao: { type: string }
        intervaloMinimo: { type: integer, minimum: 1 }
        dosagem: { type: string }
        notas: { type: string }
    DadosPontual:
      type: object
      required: [nomeResponsavel, justificacao, dataAutorizacao]
      properties:
        nomeResponsavel: { type: string }
        justificacao: { type: string, minLength: 10 }
        dataAutorizacao: { type: string, format: date-time }
    PlanoCultivoInput:
      type: object
      required: [tipo, erva]
      properties:
        tipo:
          type: string
          enum: [regular, emergencia, pontual]
        erva:
          type: string
          enum: [manjericão, hortelã, alecrim, tomilho]
        estado:
          type: string
          enum: [ativo, concluído, comprometido]
          default: ativo
        dadosRegular:
          $ref: '#/components/schemas/DadosRegular'
        dadosEmergencia:
          $ref: '#/components/schemas/DadosEmergencia'
        dadosPontual:
          $ref: '#/components/schemas/DadosPontual'
    PlanoCultivo:
      allOf:
        - $ref: '#/components/schemas/PlanoCultivoInput'
        - type: object
          properties:
            _id:
              type: string
              example: 65a1b2c3d4e5f60718293a4b
            criadoPor:
              oneOf:
                - type: string
                - $ref: '#/components/schemas/Utilizador'
            dataCriacao:
              type: string
              format: date-time
...
    LotePerda:
      type: object
      properties:
        quantidade: { type: integer, minimum: 1 }
        motivo: { type: string }
        data: { type: string, format: date-time }
    LoteCultivo:
      type: object
      required: [erva, dataInicio, quantidadeInicial]
      properties:
        _id: { type: string }
        erva: { type: string }
        planoId: { type: string }
        estado:
          type: string
          enum: [ativo, concluído, comprometido]
        dataInicio: { type: string, format: date-time }
        dataFim: { type: string, format: date-time }
        quantidadeInicial: { type: integer, minimum: 1 }
        quantidadeAtual: { type: integer, minimum: 0 }
        perdas:
          type: array
          items:
            $ref: '#/components/schemas/LotePerda'
        notas: { type: string }
        criadoPor:
          oneOf:
            - type: string
            - $ref: '#/components/schemas/Utilizador'
        dataCriacao: { type: string, format: date-time }
...
    MedicaoAmbiental:
      type: object
      required: [loteId, temperatura, humidade, luminosidade]
      properties:
        _id: { type: string }
        loteId: { type: string }
        temperatura: { type: number }
        humidade: { type: number, minimum: 0, maximum: 100 }
        luminosidade: { type: number, minimum: 0 }
        dataHora: { type: string, format: date-time }
        registadoPor:
          oneOf:
            - type: string
            - $ref: '#/components/schemas/Utilizador'
    Alerta:
      type: object
      required: [mensagem]
      properties:
        _id: { type: string }
        loteId: { type: string }
        medicaoId: { type: string }
        classificacao:
          type: string
          enum: [Informativo, Aviso, Crítico]
        mensagem: { type: string }
        estado:
          type: string
          enum: [ativo, resolvido, ignorado]
        justificacaoIgnorar: { type: string }
        resolvidoPor:
          oneOf:
            - type: string
            - $ref: '#/components/schemas/Utilizador'
        dataRegisto: { type: string, format: date-time }
        dataResolucao: { type: string, format: date-time }
...